Add files via upload
Developed a Web service, and web application. Takes the URL of an XML (.xml) file and the URL of the corresponding XMLS
(.xsd) file as input and validates the XML file against the corresponding XMLS (XSD) file. Web operation “XPathSearch” takes the URL of an XML (.xml) file and a path expression as input. It returns the path expression value of the given path.